+++ com.unity.render-pipelines.high-definition/Editor/Lighting/SerializedHDLight.cs
@@ -39,6 +39,14 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             public SerializedProperty sunDiskSize;
             public SerializedProperty sunHaloSize;
             public SerializedProperty areaLightCookie;   // We can't use default light cookies because the cookie gets reset by some safety measure on C++ side... :/
+            public SerializedProperty areaLightShadowCone;
+#if ENABLE_RAYTRACING
+            public SerializedProperty useRayTracedShadows;
+#endif
+            public SerializedProperty evsmExponent;
+            public SerializedProperty evsmLightLeakBias;
+            public SerializedProperty evsmVarianceBias;
+            public SerializedProperty evsmBlurPasses;
 
             // Improved moment shadows data
             public SerializedProperty lightAngle;
@@ -124,6 +132,14 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                     sunDiskSize = o.Find(x => x.sunDiskSize),
                     sunHaloSize = o.Find(x => x.sunHaloSize),
                     areaLightCookie = o.Find(x => x.areaLightCookie),
+                    areaLightShadowCone = o.Find(x => x.areaLightShadowCone),
+#if ENABLE_RAYTRACING
+                    useRayTracedShadows = o.Find(x => x.useRayTracedShadows),
+#endif
+                    evsmExponent = o.Find(x => x.evsmExponent),
+                    evsmVarianceBias = o.Find(x => x.evsmVarianceBias),
+                    evsmLightLeakBias = o.Find(x => x.evsmLightLeakBias),
+                    evsmBlurPasses = o.Find(x => x.evsmBlurPasses),
 
                     // Moment light
                     lightAngle = o.Find(x => x.lightAngle),